Move data catalog files to feature area folder
