+++ src/data-catalog/create-dataset/CreateDataset.test.ts
@@ -1,7 +1,7 @@
 import { screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { h } from 'react-hyperscript-helpers';
-import { CreateDataset } from 'src/pages/library/data-catalog/CreateDataset/CreateDataset';
+import { CreateDataset } from 'src/data-catalog/create-dataset/CreateDataset';
 import { renderWithAppContexts as render } from 'src/testing/test-utils';
 
 type MarkdownExports = typeof import('src/components/markdown');
