+++ src/data-catalog/create-dataset/custom-inputs/CountsInput.test.ts
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { h } from 'react-hyperscript-helpers';
-import { CountsInput } from 'src/pages/library/data-catalog/CreateDataset/CustomInputs/CountsInput';
+import { CountsInput } from 'src/data-catalog/create-dataset/custom-inputs/CountsInput';
 
 describe('CountsInput', () => {
   it('Renders a CountsInput with all fields', () => {
