+++ src/data-catalog/create-dataset/custom-inputs/CountsInput.ts
@@ -1,9 +1,9 @@
 import { div, h } from 'react-hyperscript-helpers';
-import { FormLabel } from 'src/libs/forms';
 import {
   CatalogNumberInput,
   generateIndividualInputPropsForObjectField,
-} from 'src/pages/library/data-catalog/CreateDataset/CreateDatasetInputs';
+} from 'src/data-catalog/create-dataset/CreateDatasetInputs';
+import { FormLabel } from 'src/libs/forms';
 
 export const CountsInput = ({ title, wrapperProps = {}, onChange, counts }) => {
   return div(wrapperProps, [
