+++ src/data-catalog/create-dataset/custom-inputs/PublicationInput.test.ts
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { h } from 'react-hyperscript-helpers';
-import { PublicationInput } from 'src/pages/library/data-catalog/CreateDataset/CustomInputs/PublicationInput';
+import { PublicationInput } from 'src/data-catalog/create-dataset/custom-inputs/PublicationInput';
 
 describe('PublicationInput', () => {
   it('Renders a PublicationInput with all fields and url validation', () => {
