+++ src/data-catalog/create-dataset/custom-inputs/SamplesInput.test.ts
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { h } from 'react-hyperscript-helpers';
-import { SamplesInput } from 'src/pages/library/data-catalog/CreateDataset/CustomInputs/SamplesInput';
+import { SamplesInput } from 'src/data-catalog/create-dataset/custom-inputs/SamplesInput';
 
 describe('SamplesInput', () => {
   it('Renders a SamplesInput with all fields', () => {
