+++ src/data-catalog/create-dataset/custom-inputs/SamplesInput.ts
@@ -1,8 +1,8 @@
 import * as _ from 'lodash/fp';
 import { div, h } from 'react-hyperscript-helpers';
+import { ListInput, StringInput } from 'src/data-catalog/create-dataset/CreateDatasetInputs';
 import { Samples } from 'src/libs/ajax/Catalog';
 import { FormLabel } from 'src/libs/forms';
-import { ListInput, StringInput } from 'src/pages/library/data-catalog/CreateDataset/CreateDatasetInputs';
 
 export interface SamplesInputProps {
   title?: string;
