+++ src/data-catalog/data-browser-utils.test.js
@@ -1,7 +1,5 @@
 import { render } from '@testing-library/react';
 import * as _ from 'lodash/fp';
-import { brands } from 'src/libs/brands';
-import { dataCatalogStore } from 'src/libs/state';
 import {
   datarepoSnapshotUrlFragment,
   datasetAccessTypes,
@@ -13,8 +11,10 @@ import {
   makeDatasetReleasePolicyDisplayInformation,
   prepareDatasetsForDisplay,
   workspaceUrlFragment,
-} from 'src/pages/library/dataBrowser-utils';
-import { TEST_DATASET_ONE } from 'src/pages/library/test-datasets';
+} from 'src/data-catalog/data-browser-utils';
+import { TEST_DATASET_ONE } from 'src/data-catalog/test-datasets';
+import { brands } from 'src/libs/brands';
+import { dataCatalogStore } from 'src/libs/state';
 
 beforeEach(() => {
   dataCatalogStore.reset();
