+++ src/libs/routes.js
@@ -3,6 +3,8 @@ import { compile, pathToRegexp } from 'path-to-regexp';
 import * as Analysis from 'src/analysis/Analyses';
 import * as AnalysisLauncher from 'src/analysis/AnalysisLauncher';
 import * as AppLauncher from 'src/analysis/AppLauncher';
+import * as DataBrowserDetails from 'src/data-catalog/DataBrowserDetails';
+import * as DataBrowserPreview from 'src/data-catalog/DataBrowserPreview';
 import * as ImportDataPage from 'src/import-data/ImportDataPage';
 import { routeHandlersStore } from 'src/libs/state';
 import * as AzurePreview from 'src/pages/AzurePreview';
@@ -14,9 +16,7 @@ import * as Groups from 'src/pages/groups/List';
 import * as ImportWorkflow from 'src/pages/ImportWorkflow/ImportWorkflow';
 import * as LandingPage from 'src/pages/LandingPage';
 import * as Code from 'src/pages/library/Code';
-import * as CreateDataset from 'src/pages/library/data-catalog/CreateDataset/CreateDataset';
-import * as DataBrowserDetails from 'src/pages/library/DataBrowserDetails';
-import * as DataBrowserPreview from 'src/pages/library/DataBrowserPreview';
+import * as CreateDataset from 'src/pages/library/data-catalog/create-dataset/CreateDataset';
 import * as DatasetBuilder from 'src/pages/library/datasetBuilder/DatasetBuilder';
 import * as DatasetBuilderDetails from 'src/pages/library/datasetBuilder/DatasetBuilderDetails';
 import * as Datasets from 'src/pages/library/Datasets';
